+++ src/Faker/Provider/Html.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Vtec\Crud\Faker\Provider;
+namespace Okami101\LaravelVuetifyAdmin\Faker\Provider;
 
 use Faker\Generator;
 use Faker\Provider\Base;